feat(TTB-043): Renamed Description.jsx
